[STS3032] Fix bugs, it works!
